requirements.txt
Flask==2.2.2
Flask_APScheduler==1.12.4
Flask_Login==0.6.2
flask_sqlalchemy==3.0.2
pandas==1.5.1
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.42
Werkzeug==2.2.2
